+++ eapi-editor/src/pages/project/tags/tags.vue
@@ -3,7 +3,7 @@
     <Form ref="formInline" inline class="project-tags-form">
       <FormItem>
         <!--<Button type="primary" @click="handleSubmit('formInline')">Signin</Button>-->
-        <Button type="primary" icon="edit" @click="newTag">新建标签</Button>
+        <Button type="primary" icon="md-add" @click="newTag">新建标签</Button>
       </FormItem>
       <FormItem prop="search">
         <Input type="text" placeholder="搜索"/>
@@ -156,15 +156,24 @@ export default {
       this.addTagModal = true;
     },
     deleteTags() {
-      deleteTagInBatch(this.selection, (response) => {
-        if (response.header.code === '0') {
-          this.$Message.success('删除成功！');
-          this.showEditMenus = false;
-          this.init();
-        } else {
-          this.$Message.error(response.header.message);
-        }
-      });
+			this.$Modal.confirm({
+				title: '确认删除？',
+				content: '<p>删除数据不可恢复！</p><p>确认要删除吗？</p>',
+				onOk: () => {
+					deleteTagInBatch(this.selection, (response) => {
+						if (response.header.code === '0') {
+							this.$Message.success('删除成功！');
+							this.showEditMenus = false;
+							this.init();
+						} else {
+							this.$Message.error(response.header.message);
+						}
+					});
+				},
+				onCancel: () => {
+					this.$Message.success("取消删除！");
+				}
+			});
     }
   },
   mounted() {
